+++ game/client/c_basecombatcharacter.h
@@ -84,8 +84,8 @@ public:
 	// This is a sort of hack back-door only used by physgun!
 	void SetAmmoCount( int iCount, int iAmmoIndex );
 
-	float				GetNextAttack() const { return m_flNextAttack; }
-	void				SetNextAttack( float flWait ) { m_flNextAttack = flWait; }
+	double				GetNextAttack() const { return m_flNextAttack; }
+	void				SetNextAttack( double flWait ) { m_flNextAttack = flWait; }
 
 	virtual int			BloodColor();
 
@@ -105,7 +105,7 @@ public:
 
 public:
 
-	float			m_flNextAttack;
+	double			m_flNextAttack;
 
 protected:
